refactor(tenant-api): extract PR-mode write-back into gitops/writer_pr.go
Continue the writer.go decomposition (Cycle 4 gitcmd.go, Cycle 5 writer_special.go):
move the PR-mode write-back block — PRWriteResult, WritePR, WritePRBatch, PRBatchOp —
into a new same-package file writer_pr.go, so the feature-branch/PR flow reads apart
from the direct commit-on-write path. writer.go drops 757 -> 552 lines (999 -> 552
across the three splits). The base-branch anchoring helpers and shared commit /
admission primitives stay in writer.go (same package, still resolvable).

Pure intra-package file move: same package gitops, no caller/test/signature/import
changes. Adversarial review confirms the 4-symbol set is conserved (declared once),
the WritePR rollback paths + #641 pushed-gated branch cleanup are byte-identical,
and handler call sites resolve. go build / vet / gitops + handler PR/Batch tests green.

Self-Review-Pass-2: adversarial: faithful move; 4 symbols conserved; rollback byte-identical
